requirements.txt
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
attrs @ file:///croot/attrs_1695717823297/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4==4.12.3
bleach==6.1.0
Bottleneck @ file:///opt/conda/conda-bld/bottleneck_1657175564434/work
Brotli @ file:///tmp/abs_ecyw11_7ze/croots/recipe/brotli-split_1659616059936/work
certifi @ file:///croot/certifi_1707229174982/work/certifi
cffi @ file:///tmp/abs_98z5h56wf8/croots/recipe/cffi_1659598650955/work
charset-normalizer==3.1.0
cloudpickle @ file:///croot/cloudpickle_1683040006038/work
comm @ file:///croot/comm_1671231121260/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
Cython @ file:///croot/cython_1690297244722/work
dcor @ file:///home/conda/feedstock_root/build_artifacts/dcor_1672090954146/work
debugpy @ file:///croot/debugpy_1690905042057/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml==0.7.1
entrypoints @ file:///tmp/build/80754af9/entrypoints_1649908313000/work
et-xmlfile==1.1.0
exceptiongroup @ file:///croot/exceptiongroup_1668714342571/work
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
Farama-Notifications @ file:///croot/farama-notifications_1683058702908/work
fastjsonschema @ file:///opt/conda/conda-bld/python-fastjsonschema_1661371079312/work
fdasrsf @ file:///home/conda/feedstock_root/build_artifacts/fdasrsf_1692668236999/work
findiff @ file:///home/conda/feedstock_root/build_artifacts/findiff_1692974147098/work
fonttools==4.25.0
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455533097/work
GPUtil @ file:///home/conda/feedstock_root/build_artifacts/gputil_1590646865081/work
gymnasium @ file:///croot/gymnasium-split_1683292852344/work
idna==3.4
import-ipynb==0.1.4
inflate64 @ file:///croot/inflate64_1683829557981/work
ipykernel @ file:///croot/ipykernel_1691121631942/work
ipympl @ file:///home/conda/feedstock_root/build_artifacts/ipympl_1676535632179/work
ipynb==0.5.1
ipython @ file:///croot/ipython_1694181358621/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets @ file:///croot/ipywidgets_1679394798311/work
jax-jumpy @ file:///croot/jax-jumpy_1683230527192/work
jedi @ file:///tmp/build/80754af9/jedi_1644315229345/work
Jinja2==3.1.3
joblib @ file:///tmp/build/80754af9/joblib_1635411271373/work
jsonschema @ file:///croot/jsonschema_1676558650973/work
jupyter-client @ file:///opt/conda/conda-bld/jupyter_client_1643638337975/work
jupyter_core @ file:///croot/jupyter_core_1679906564508/work
jupyterlab-widgets @ file:///croot/jupyterlab_widgets_1679055282532/work
jupyterlab_pygments==0.3.0
kiwisolver @ file:///opt/conda/conda-bld/kiwisolver_1653292039266/work
lazy_loader @ file:///croot/lazy_loader_1695850097191/work
llvmlite @ file:///croot/llvmlite_1683555126642/work
MarkupSafe==2.1.5
matplotlib @ file:///tmp/build/80754af9/matplotlib-suite_1647441664166/work
matplotlib-inline @ file:///opt/conda/conda-bld/matplotlib-inline_1662014470464/work
mistune==3.0.2
mkl-fft==1.3.1
mkl-random @ file:///home/builder/ci_310/mkl_random_1641843545607/work
mkl-service==2.4.0
mpmath @ file:///croot/mpmath_1690848262763/work
multimethod @ file:///croot/multimethod_1687530375480/work
multivolumefile @ file:///tmp/build/80754af9/multivolumefile_1633093904359/work
munkres==1.1.4
nbclient==0.9.0
nbconvert==7.16.2
nbformat @ file:///croot/nbformat_1694616755618/work
nbimporter==0.3.4
nest-asyncio @ file:///croot/nest-asyncio_1672387112409/work
numba @ file:///home/conda/feedstock_root/build_artifacts/numba_1687804769779/work
numexpr @ file:///opt/conda/conda-bld/numexpr_1656940300424/work
numpy==1.22.4
openpyxl==3.0.10
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.4.3
pandocfilters==1.5.1
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
patsy==0.5.2
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.2.0
platformdirs @ file:///croot/platformdirs_1692205439124/work
ply==3.11
prompt-toolkit @ file:///croot/prompt-toolkit_1672387306916/work
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
py7zr @ file:///croot/py7zr_1684257643914/work
pybcj @ file:///croot/pybcj_1683831074238/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodomex @ file:///opt/conda/conda-bld/pycryptodomex_1661450193025/work
Pygments @ file:///croot/pygments_1684279966437/work
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
pyppmd @ file:///croot/pyppmd_1683831074175/work
PyQt5-sip==12.11.0
pyrsistent @ file:///home/builder/ci_310/pyrsistent_1640807196327/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz @ file:///opt/conda/conda-bld/pytz_1654762638606/work
PyYAML @ file:///croot/pyyaml_1698096049011/work
pyzmq @ file:///croot/pyzmq_1686601365461/work
pyzstd==0.15.0
rdata @ file:///home/conda/feedstock_root/build_artifacts/rdata_1654609548457/work
requests==2.28.2
scikit-datasets @ file:///home/conda/feedstock_root/build_artifacts/scikit-datasets_1691517780458/work
scikit-fda @ file:///home/conda/feedstock_root/build_artifacts/scikit-fda_1674768509351/work
scikit-learn @ file:///croot/scikit-learn_1684781056615/work
scipy @ file:///home/builder/ci_310/scipy_1641389939032/work
seaborn @ file:///tmp/build/80754af9/seaborn_1629307859561/work
sip @ file:///tmp/abs_44cd77b_pu/croots/recipe/sip_1659012365470/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
soupsieve==2.5
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
statsmodels @ file:///tmp/build/80754af9/statsmodels_1648015170016/work
sympy @ file:///croot/sympy_1668202399572/work
texttable @ file:///tmp/build/80754af9/texttable_1633095311315/work
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
tinycss2==1.2.1
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
torch==1.11.0+cu113
torchvision==0.12.0+cu113
tornado @ file:///home/builder/ci_310/tornado_1640795392144/work
tqdm @ file:///croot/tqdm_1679561862951/work
traitlets @ file:///croot/traitlets_1671143879854/work
typing_extensions @ file:///croot/typing_extensions_1690297465030/work
urllib3==1.26.15
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings==0.5.1
widgetsnbextension @ file:///croot/widgetsnbextension_1679313860248/work
xarray @ file:///croot/xarray_1689041471350/work
